Refactor JobItemExecutorFactoryTest fixture
