+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/executor/type/fixture/job/DetailedFooJob.java
@@ -15,9 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.type.fixture;
+package org.apache.shardingsphere.elasticjob.lite.executor.type.fixture.job;
 
-import org.apache.shardingsphere.elasticjob.lite.api.job.ElasticJob;
-
-public interface FooJob extends ElasticJob {
+public final class DetailedFooJob implements FooJob {
 }
