+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/executor/type/fixture/job/FailedJob.java
@@ -15,21 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.type.fixture;
+package org.apache.shardingsphere.elasticjob.lite.executor.type.fixture.job;
 
-import org.apache.shardingsphere.elasticjob.lite.api.job.ShardingContext;
-import org.apache.shardingsphere.elasticjob.lite.api.job.config.JobConfiguration;
-import org.apache.shardingsphere.elasticjob.lite.executor.JobFacade;
-import org.apache.shardingsphere.elasticjob.lite.executor.type.JobItemExecutor;
+import org.apache.shardingsphere.elasticjob.lite.api.job.ElasticJob;
 
-public final class FooJobExecutor implements JobItemExecutor<FooJob> {
-    
-    @Override
-    public void process(final FooJob elasticJob, final JobConfiguration jobConfig, final JobFacade jobFacade, final ShardingContext shardingContext) {
-    }
-    
-    @Override
-    public Class<FooJob> getElasticJobClass() {
-        return FooJob.class;
-    }
+public final class FailedJob implements ElasticJob {
 }
